requirements.txt
aioredis==1.3.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
beautifulsoup4==4.13.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.115.11
fastapi-cache2==0.2.2
feedparser==6.0.11
filelock==3.17.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.29.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
packaging==24.2
pendulum==3.0.0
pluggy==1.5.0
preshed==3.0.9
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.46.0
sympy==1.13.3
tenacity==8.2.3
textblob==0.17.1
thinc==8.2.5
time-machine==2.16.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.67.1
transformers==4.38.2
tweepy==4.15.0
typer==0.9.4
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.3.4
Wikipedia-API==0.8.1
